+++ lib/Data/Time/Calendar.hs
@@ -1,9 +1,11 @@
 module Data.Time.Calendar
 (
     module Data.Time.Calendar.Days,
-    module Data.Time.Calendar.Gregorian
+    module Data.Time.Calendar.Gregorian,
+    module Data.Time.Calendar.Week
 ) where
 
 import Data.Time.Calendar.Days
 import Data.Time.Calendar.Gregorian
+import Data.Time.Calendar.Week
 import Data.Time.Format()